refactor(redskilled): isolate GitHub budget projection Refs #3387
